criar e deletar mensages

+++ base/views.py
@@ -1,9 +1,12 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
+from django.http import HttpResponse
 from django.contrib.auth.models import User
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth import authenticate, login, logout
-from .models import Room, Topic
+from django.contrib.auth.forms import UserCreationForm
+from .models import Room, Topic, Message
 from .forms import RoomForm
 # Create your views here.
 # rooms = [
@@ -13,9 +16,14 @@ from .forms import RoomForm
 # ]
 
 def loginPage(request):
+    
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
 
     if request.method == 'POST':
-        username= request.POST.get('username')
+        username= request.POST.get('username').lower()
         password= request.POST.get('password')
         
         try:
@@ -32,7 +40,7 @@ def loginPage(request):
         else:
             messages.error(request, 'Usuário ou Senha não existem')
 
-    context = {}
+    context = {'page': page}
 
     return render(request,'base/login_register.html',context)
 
@@ -40,6 +48,23 @@ def logoutUser(request):
     logout(request)
     return redirect('home')
 
+
+def registerPage(request):
+    form = UserCreationForm()
+    
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request,'Um erro ocorreu durante o registro')
+
+
+    return render(request, 'base/login_register.html', {'form': form})
 
 def home(request):
     
@@ -64,12 +89,32 @@ def home(request):
 
 def room(request,pk):
     room = Room.objects.get(id=pk)
-    context = {'room': room}
+    room_messages = room.message_set.all().order_by('-created')
+    participants = room.participants.all()
+
+    if request.method =='POST':
+        message = Message.objects.create(
+            user=request.user,
+            room=room,
+            body=request.POST.get('body')
+        )  
+
+        room.participants.add(request.user)
+        return redirect ('room', pk=room.id)
+
+    context = {
+        'room': room,
+        'room_messages': room_messages,  
+        'participants': participants,  
+    }
 
     return render(request, 'base/room.html', context)
 
+
+@login_required(login_url='login')
 def createRoom(request):
 
+    
     form = RoomForm
     if request.method == 'POST':
         form = RoomForm(request.POST)
@@ -82,10 +127,14 @@ def createRoom(request):
 
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
 
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("Você não tem permissão para estar aqui.")
 
     if request.method =='POST':
         form = RoomForm(request.POST,instance=room)
@@ -100,13 +149,37 @@ def updateRoom(request,pk):
 
     return render(request,'base/room_form.html', context) 
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
 
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("Você não tem permissão para estar aqui.")
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
 
     context = {'obj':room}
+    
+    return render(request, 'base/delete.html', context)
+
+
+
+
+@login_required(login_url='login')
+def deleteMessage(request, pk):
+
+    message = Message.objects.get(id=pk)
+
+    if request.user != message.user:
+        return HttpResponse("Você não tem permissão para estar aqui.")
+
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+
+    context = {'obj':message}
     
     return render(request, 'base/delete.html', context)
